requirements.txt
transformers>=4.25.1
beautifulsoup4>=4.11.1
lxml>=4.9.1
pyyaml>=6.0
datasets>=2.8.0
python-dateutil>=2.8.2
tqdm>=4.64.1